Module Base_quickcheck

module Generator : sig ... end

Generators are sources of random values. Every randomized test needs a generator to produce its inputs.

module Observer : sig ... end

Observers create random functions. Generator.fn creates a random function using an observer for the input type and a generator for the output type.

module Shrinker : sig ... end

Shrinkers produce small values from large values. When a random test case fails, a shrinker finds the simplest version of the problem.

module Test : sig ... end

Use the Test module to run randomized tests. Each randomized test needs a generator, a shrinker, and a property to test.

module Export : sig ... end

Provides default generators, observers, and shrinkers for built-in types. Follows ppx_quickcheck naming conventions.

include Export

val quickcheck_generator_unit : Base.unit Generator.t

val quickcheck_generator_bool : Base.bool Generator.t

val quickcheck_generator_char : Base.char Generator.t

val quickcheck_generator_string : Base.string Generator.t

val quickcheck_generator_int : Base.int Generator.t

val quickcheck_generator_int32 : Base.int32 Generator.t

val quickcheck_generator_int64 : Base.int64 Generator.t

val quickcheck_generator_nativeint : Base.nativeint Generator.t

val quickcheck_generator_float : Base.float Generator.t

val quickcheck_observer_unit : Base.unit Observer.t

val quickcheck_observer_bool : Base.bool Observer.t

val quickcheck_observer_char : Base.char Observer.t

val quickcheck_observer_string : Base.string Observer.t

val quickcheck_observer_int : Base.int Observer.t

val quickcheck_observer_int32 : Base.int32 Observer.t

val quickcheck_observer_int64 : Base.int64 Observer.t

val quickcheck_observer_nativeint : Base.nativeint Observer.t

val quickcheck_observer_float : Base.float Observer.t

val quickcheck_shrinker_unit : Base.unit Shrinker.t

val quickcheck_shrinker_bool : Base.bool Shrinker.t

val quickcheck_shrinker_char : Base.char Shrinker.t

val quickcheck_shrinker_string : Base.string Shrinker.t

val quickcheck_shrinker_int : Base.int Shrinker.t

val quickcheck_shrinker_int32 : Base.int32 Shrinker.t

val quickcheck_shrinker_int64 : Base.int64 Shrinker.t

val quickcheck_shrinker_nativeint : Base.nativeint Shrinker.t

val quickcheck_shrinker_float : Base.float Shrinker.t

val quickcheck_generator_option : 'a Generator.t -> 'a Base.option Generator.t

val quickcheck_generator_list : 'a Generator.t -> 'a Base.list Generator.t

val quickcheck_observer_option : 'a Observer.t -> 'a Base.option Observer.t

val quickcheck_observer_list : 'a Observer.t -> 'a Base.list Observer.t

val quickcheck_shrinker_option : 'a Shrinker.t -> 'a Base.option Shrinker.t

val quickcheck_shrinker_list : 'a Shrinker.t -> 'a Base.list Shrinker.t