Module Base_quickcheck.Generator

Generators are sources of random values. Every randomized test needs a generator to produce its inputs.

type +('a : any) t

Basic Generators

These are good default generators for tests over types from OCaml and Base. They are designed to hit corner cases reasonably often, and also generate reasonably good coverage of common cases and arbitrary values.

val unit : Base.unit t

val bool : Base.bool t

val char : Base.char t

val string : Base.string t

val bytes : Base.bytes t

val int : Base.int t

val int32 : Base.int32 t

val int63 : Base.Int63.t t

val int64 : Base.int64 t

val nativeint : Base.nativeint t

val float : Base.float t

val sexp : Base.Sexp.t t

val option : 'a. 'a t -> 'a Base.option t

val or_null : 'a t -> 'a Base.or_null t

val list : 'a. 'a t -> 'a Base.list t

val array : 'a t -> 'a Base.array t

val ref : 'a. 'a t -> 'a Base.ref t

val lazy_t : 'a t -> 'a Base.Lazy.t t

val either : 'a 'b. 'a t -> 'b t -> ('a, 'b) Base.Either.t t

val result : 'a 'b. 'a t -> 'b t -> ('a, 'b) Base.Result.t t

This helper module type exists separately just to open Bigarray in its scope.

val bigstring : (Base.char, Bigarray.int8_unsigned_elt, Bigarray.c_layout) Bigarray.Array1.t t

val float32_vec : (Base.float, Bigarray.float32_elt, Bigarray.fortran_layout) Bigarray.Array1.t t

val float64_vec : (Base.float, Bigarray.float64_elt, Bigarray.fortran_layout) Bigarray.Array1.t t

val float32_mat : (Base.float, Bigarray.float32_elt, Bigarray.fortran_layout) Bigarray.Array2.t t

val float64_mat : (Base.float, Bigarray.float64_elt, Bigarray.fortran_layout) Bigarray.Array2.t t

val fn : 'a 'b. 'a Base_quickcheck__.Observer0.t -> 'b t -> ('a -> 'b) t @@ portable

Generates random functions that use the given observer to perturb the pseudo-random state that is then used to generate the output value. The resulting functions are therefore deterministic, assuming the observer is deterministic.

val map_t_m : 'key 'data 'cmp. ('key, 'cmp) Base.Comparator.Module.t -> 'key t -> 'data t -> ('key, 'data, 'cmp) Base.Map.t t @@ portable

val set_t_m : ('elt, 'cmp) Base.Comparator.Module.t -> 'elt t -> ('elt, 'cmp) Base.Set.t t @@ portable

val map_tree_using_comparator : 'key 'data 'cmp. comparator:('key, 'cmp) Base.Comparator.t -> 'key t -> 'data t -> ('key, 'data, 'cmp) Base.Map.Using_comparator.Tree.t t @@ portable

val set_tree_using_comparator : comparator:('elt, 'cmp) Base.Comparator.t -> 'elt t -> ('elt, 'cmp) Base.Set.Using_comparator.Tree.t t @@ portable

Combining and Modifying Generators

val of_list : 'a. 'a Base.list -> 'a t @@ portable

Produces any of the given values, weighted uniformly.

val union : 'a. 'a t Base.list -> 'a t @@ portable

Chooses among the given generators, weighted uniformly; then chooses a value from that generator.

include Base.Applicative.S__'value_or_null_mod_maybe_null' with type ('a : value_or_null) t := 'a t

val apply : 'a 'b 'p 'q. ('a -> 'b) t -> 'a t -> 'b t

Applies the functions in one t to the values in another. Well-behaved applicatives satisfy these "laws", using <*> as infix apply:

return Fn.id <*> t is equivalent to t

return Fn.compose <*> tf <*> tg <*> tx is equivalent to tf <*> (tg <*> tx)

return f <*> return x is equivalent to return (f x)

tf <*> return x is equivalent to return (fun f -> f x) <*> tf

val both : 'a 'b 'p 'q. 'a t -> 'b t -> ('a * 'b) t

Combines values in two ts as tuples. Using <*> as infix apply, equivalent to return (fun a b -> a, b) <*> ta <*> tb.

val map2 : 'a 'b 'c 'p 'q. 'a t -> 'b t -> f:('a -> 'b -> 'c) -> 'c t

Combines the contents of two ts. Using <*> as infix apply, equivalent to return f <*> ta <*> tb.

val map3 : 'a 'b 'c 'd 'p 'q. 'a t -> 'b t -> 'c t -> f:('a -> 'b -> 'c -> 'd) -> 'd t

Combines the contents of three ts. Using <*> as infix apply, equivalent to return f <*> ta <*> tb <*> tc.

val (<*>) : 'a 'b 'p 'q. ('a -> 'b) t -> 'a t -> 'b t

val (<*) : 'a 'p 'q. 'a t -> unit t -> 'a t

val (*>) : 'a 'p 'q. unit t -> 'a t -> 'a t

module Applicative_infix : sig ... end

include Base.Monad.S__'value_or_null_mod_maybe_null' with type ('a : value_or_null) t := 'a t

val return : 'a 'i 'p 'q. 'a -> 'a t

Convert a value to a t.

val map : 'a 'b 'i 'j 'p 'q. 'a t -> f:('a -> 'b) -> 'b t

Transforms the contents of a t.

val bind : 'a 'b 'i 'j 'k 'p 'q. 'a t -> f:('a -> 'b t) -> 'b t

Sequences computations. bind t ~f computes f v for value(s) v in t. Well-behaved monads satisfy these "laws" (where ( >>= ) is the infix bind operator):

map t ~f is equivalent to bind t ~f:(fun x -> return (f x))

return x >>= f is equivalent to f x

t >>= return is equivalent to t

(t >>= f) >>= g is equivalent to t >>= fun x -> f x >>= g

val join : 'a 'i 'j 'k 'p 'q. 'a t t -> 'a t

Combines nested t into just one layer. Equivalent to bind t ~f:Fn.id.

val ignore_m : 'a 'i 'j 'p 'q. 'a t -> unit t

Ignores contained values of t. Equivalent to map t ~f:ignore.

val all : 'a 'i 'p 'q. 'a t list -> 'a list t

Combines a list of t.

val all_unit : 'i 'p 'q. unit t list -> unit t

Combines a list of t whose contents are unimportant.

val (>>=) : 'a 'b 'i 'j 'k 'p 'q. 'a t -> ('a -> 'b t) -> 'b t

Infix bind.

val (>>|) : 'a 'b 'i 'j 'p 'q. 'a t -> ('a -> 'b) -> 'b t

Infix map.

module Monad_infix : sig ... end

module Let_syntax : sig ... end

module Portable : sig ... end

Size of Random Values

Base_quickcheck threads a size parameter through generators to limit the size of unbounded types. Users of Base_quickcheck often do not need to think about the size parameter; the default generators handle it sensibly. Generators of atomic types ignore it, generators of bounded-size containers like both and either thread it through unchanged, and generators of unbounded-size containers like list and set_t_m distribute the size they are given among their constituents.

The bindings below allow direct manipulation of the size parameter in cases where users want a custom treatment of sizes. There is no prescribed meaning of the size parameter for any given type other than that it must be non-negative. As a general guideline, however, the time and space used to generate a value should be proportional to the size parameter at most.

The size parameter should be treated as an upper bound but not as a lower bound, so for example a generator given a size parameter of 2 should have a chance to generate values of size 0 or 1 as well. If the size parameter is treated as a lower bound, then for example members of tuples will always be generated at the same size, and test cases for members of different size will not be covered.

val size : Base.int t @@ portable

Returns the current size parameter.

val with_size : 'a. 'a t -> size:Base.int -> 'a t @@ portable

Produces a generator that ignores the size parameter passed in by Base_quickcheck and instead uses the given ~size argument. Most often used with size to reduce the size when dispatching to generators for subparts of a value.

For example, here is a use of with_size and size to create a generator for optional lists. We are careful to generate None even at non-zero sizes; see the note above about not using size as a lower bound.

  let optional_list generator =
    let open Let_syntax in
    match%bind both size bool with
    | 0, _ | _, false -> return None
    | k, _ ->
      let%map elements = with_size ~size:(k - 1) (list generator) in
      Some elements
  ;;

val sizes : ?min_length:Base.int -> ?max_length:Base.int -> Base.unit -> Base.int Base.list t @ portable @@ portable

Produces a list of sizes that distribute the current size among list elements. The min_length and max_length parameters can be used to bound the length of the result.

This is the distribution used by generators such as list to divide up size among elements.

This function is designed so that elements of list are always generated at strictly smaller size than the list itself. The technical invariant is: if size_list is generated by with_size ~size:n (sizes ~min_length ()), then:

  List.length size_list - min_length + List.sum (module Int) size_list <= n

Filtering Generators

val filter : 'a. 'a t -> f:('a -> Base.bool) -> 'a t @@ portable

Produces values for which f returns true. If f returns false, retries with size incremented by 1. This avoids filter getting stuck if all values at a given size fail f; see the note above about not using size as a lower bound.

val filter_map : 'a 'b. 'a t -> f:('a -> 'b Base.option) -> 'b t @@ portable

When f produces Some x, produces x. If f returns None, retries with size incremented by 1, as with filter.

Generating Recursive Values

val recursive_union : 'a. 'a t Base.list -> f:('a t -> 'a t Base.list) -> 'a t @@ portable

Ties the recursive knot to produce generators for recursive types that have multiple clauses, separating base cases from recursive cases. At size 0, only base cases are produced; at size n > 0, the base cases are produced at size n along with the recursive cases at size n-1. Raises if the list of base cases is empty or if the list of recursive cases is empty.

For example, here is a use of recursive_union to create a generator for an expression datatype.

  type exp =
    | Int of int
    | Bool of bool
    | If of exp * exp * exp
    | Add of exp * exp

  let exp_generator =
    recursive_union
      [ map int ~f:(fun i -> Int i); map bool ~f:(fun b -> Bool b) ]
      ~f:(fun exp ->
        let open Let_syntax in
        [ (let%map a = exp
           and b = exp
           and c = exp in
           If (a, b, c))
        ; (let%map a = exp
           and b = exp in
           Add (a, b))
        ])
  ;;

val fixed_point : 'a. ('a t -> 'a t) -> 'a t @@ portable

Like recursive_union, without separate clauses or automatic size management. Useful for generating recursive types that don't fit the clause structure of recursive_union.

For example, here is a use of fixed_point to create a generator for N-ary trees. No manual size management is needed, as Generator.list guarantees to generate list elements at strictly smaller sizes than the list itself.

  type tree = Node of tree list

  let tree_generator =
    fixed_point (fun tree -> map (list tree) ~f:(fun trees -> Node trees))
  ;;

val of_lazy : 'a. 'a t Base.Lazy.t -> 'a t @@ portable

Creates a t that forces the lazy argument as necessary. Can be used to tie (mutually) recursive knots.

val of_portable_lazy : 'a. 'a t Base.Portable_lazy.t @ portable -> 'a t @ portable @@ portable

Like of_lazy, but for Portable_lazy.t.

Custom Random Distributions

val of_weighted_list : 'a. (Base.float * 'a) Base.list -> 'a t @@ portable

Produces one of the given values, chosen with the corresponding weight. Weights must be non-negative and must have a strictly positive sum.

val weighted_union : 'a. (Base.float * 'a t) Base.list -> 'a t @@ portable

Produces one of the given generators, chosen with the corresponding weight, then chooses a value from that generator. Weights must be non-negative and must have a strictly positive sum.

val weighted_recursive_union : 'a. (Base.float * 'a t) Base.list -> f:('a t -> (Base.float * 'a t) Base.list) -> 'a t @@ portable

Like recursive_union, with explicit weights for each clause. Weights must be non-negative and the recursive case weights must have a strictly positive sum.

Integer Distributions

val small_positive_or_zero_int : Base.int t @@ portable

Produces an integer between 0 and an unspecified upper bound which is proportional to size. This is a good generator to use for sizes of values like strings which have a variable number of fixed-size elements.

val small_strictly_positive_int : Base.int t @@ portable

Like small_positive_or_zero_int but with a minimum of 1.

Uniform Unbounded Distributions

These generators produce any value of the relevant integer type with uniform weight. The default generators for these types differ in that they give higher weight to corner cases, e.g. min_value and max_value.

val int_uniform : Base.int t @@ portable

val int32_uniform : Base.int32 t @@ portable

val int63_uniform : Base.Int63.t t @@ portable

val int64_uniform : Base.int64 t @@ portable

val nativeint_uniform : Base.nativeint t @@ portable

Bounded Distributions

These generators produce any value between the given inclusive bounds, which must be given in nondecreasing order. Higher weight is given to corner cases, e.g. the bounds themselves.

val int_inclusive : Base.int -> Base.int -> Base.int t @ portable @@ portable

val int32_inclusive : Base.int32 -> Base.int32 -> Base.int32 t @ portable @@ portable

val int63_inclusive : Base.Int63.t -> Base.Int63.t -> Base.Int63.t t @ portable @@ portable

val int64_inclusive : Base.int64 -> Base.int64 -> Base.int64 t @ portable @@ portable

val nativeint_inclusive : Base.nativeint -> Base.nativeint -> Base.nativeint t @ portable @@ portable

Uniform Bounded Distributions

These generators produce any value between the given inclusive bounds, which must be given in nondecreasing order. All values are given equal weight.

val int_uniform_inclusive : Base.int -> Base.int -> Base.int t @ portable @@ portable

val int32_uniform_inclusive : Base.int32 -> Base.int32 -> Base.int32 t @ portable @@ portable

val int63_uniform_inclusive : Base.Int63.t -> Base.Int63.t -> Base.Int63.t t @ portable @@ portable

val int64_uniform_inclusive : Base.int64 -> Base.int64 -> Base.int64 t @ portable @@ portable

val nativeint_uniform_inclusive : Base.nativeint -> Base.nativeint -> Base.nativeint t @ portable @@ portable

Uniform in Log Space Distributions

These generators produce any value between the given inclusive, non-negative bounds, choosing bit-length in that range uniformly and then uniformly among values with that bit-length between the bounds. The bounds must be given in nondecreasing order.

val int_log_uniform_inclusive : Base.int -> Base.int -> Base.int t @ portable @@ portable

val int32_log_uniform_inclusive : Base.int32 -> Base.int32 -> Base.int32 t @ portable @@ portable

val int63_log_uniform_inclusive : Base.Int63.t -> Base.Int63.t -> Base.Int63.t t @ portable @@ portable

val int64_log_uniform_inclusive : Base.int64 -> Base.int64 -> Base.int64 t @ portable @@ portable

val nativeint_log_uniform_inclusive : Base.nativeint -> Base.nativeint -> Base.nativeint t @ portable @@ portable

Log Space Distributions

Like the *_log_uniform_inclusive bindings above, but giving additional weight to corner cases, e.g. the given bounds.

val int_log_inclusive : Base.int -> Base.int -> Base.int t @ portable @@ portable

val int32_log_inclusive : Base.int32 -> Base.int32 -> Base.int32 t @ portable @@ portable

val int63_log_inclusive : Base.Int63.t -> Base.Int63.t -> Base.Int63.t t @ portable @@ portable

val int64_log_inclusive : Base.int64 -> Base.int64 -> Base.int64 t @ portable @@ portable

val nativeint_log_inclusive : Base.nativeint -> Base.nativeint -> Base.nativeint t @ portable @@ portable

Geometric Distributions

These generators produce a geometric distribution with a given minimum and probabilty p. In other words, with probability p, the minimum is produced. Otherwise, a value is effectively produced from a geometric distribution with the same p and a minimum one higher, although the implementation can be more efficent than this. If the result overflows, the function returns max_value for the integer type.

Raises if p <. 0. || 1. <. p..

val int_geometric : Base.int -> p:Base.float -> Base.int t @ portable @@ portable

val int32_geometric : Base.int32 -> p:Base.float -> Base.int32 t @ portable @@ portable

val int63_geometric : Base.Int63.t -> p:Base.float -> Base.Int63.t t @ portable @@ portable

val int64_geometric : Base.int64 -> p:Base.float -> Base.int64 t @ portable @@ portable

val nativeint_geometric : Base.nativeint -> p:Base.float -> Base.nativeint t @ portable @@ portable

Floating Point Distributions

val float_inclusive : Base.float -> Base.float -> Base.float t @ portable @@ portable

Generates values between the given bounds, inclusive, which must be finite and in nondecreasing order. Weighted toward boundary values.

val float_uniform_exclusive : Base.float -> Base.float -> Base.float t @ portable @@ portable

Generates values between the given bounds, exclusive, which must be finite and in increasing order, with at least one float value between them. Weighted approximately uniformly across the resulting range, rounding error notwithstanding.

val float_without_nan : Base.float t @@ portable

val float_finite : Base.float t @@ portable

val float_strictly_positive : Base.float t @@ portable

val float_strictly_negative : Base.float t @@ portable

val float_positive_or_zero : Base.float t @@ portable

val float_negative_or_zero : Base.float t @@ portable

val float_of_class : Base.Float.Class.t -> Base.float t @ portable @@ portable

Character Distributions

val char_lowercase : Base.char t @@ portable

val char_uppercase : Base.char t @@ portable

val char_digit : Base.char t @@ portable

val char_alpha : Base.char t @@ portable

val char_alphanum : Base.char t @@ portable

val char_whitespace : Base.char t @@ portable

val char_print : Base.char t @@ portable

val char_uniform_inclusive : Base.char -> Base.char -> Base.char t @ portable @@ portable

String Distributions

val string_non_empty : Base.string t @@ portable

val string_with_length : length:Base.int -> Base.string t @ portable @@ portable

val string_of : Base.char t -> Base.string t @@ portable

val string_non_empty_of : Base.char t -> Base.string t @@ portable

val string_with_length_of : Base.char t -> length:Base.int -> Base.string t @@ portable

val string_like : Base.string -> Base.string t @ portable @@ portable

Produces strings similar to the input, with some number of edits.

Sexp Distributions

val sexp_of : Base.string t -> Base.Sexp.t t @@ portable

Produces s-expressions whose atoms are chosen from the given string distribution.

List Distributions

val list_non_empty : 'a. 'a t -> 'a Base.list t @@ portable

val list_with_length : 'a. 'a t -> length:Base.int -> 'a Base.list t @@ portable

val list_filtered : 'a Base.list -> 'a Base.list t @@ portable

Randomly drops elements from a list. The length of each result is chosen uniformly between 0 and the length of the input, inclusive.

val list_permutations : 'a Base.list -> 'a Base.list t @@ portable

Produces permutations of the given list, weighted uniformly.

val fold_until : ?min_length:Base.int -> ?max_length:Base.int -> init:'acc -> f:('acc -> ('acc, 'final) Base.Continue_or_stop.t t) -> finish:('acc -> 'final) -> Base.unit -> 'final t @@ portable

Bigarray Distributions

val bigarray1 : 'a t -> ('a, 'b) Bigarray.kind -> 'c Bigarray.layout -> length:Base.int Base.option -> ('a, 'b, 'c) Bigarray.Array1.t t

val bigstring_with_length : length:Base.int -> (Base.char, Bigarray.int8_unsigned_elt, Bigarray.c_layout) Bigarray.Array1.t t @ portable

val float32_vec_with_length : length:Base.int -> (Base.float, Bigarray.float32_elt, Bigarray.fortran_layout) Bigarray.Array1.t t @ portable

val float64_vec_with_length : length:Base.int -> (Base.float, Bigarray.float64_elt, Bigarray.fortran_layout) Bigarray.Array1.t t @ portable

Low-Level Interface

These functions provide direct access to the pseudo-random state threaded through Base_quickcheck generators. Most users should not need these functions.

val perturb : 'a. 'a t -> Base.int -> 'a t @@ portable

Passes in additional "salt" used to perturb the pseudo-random state used to generate random values. Generators' output is intended to be deterministic for any initial pseudorandom state, so perturb can be used to generate a new generator with the same distribution that nonetheless produces different values from the original for any given pseudo-random state.

val create : 'a. (size:Base.int -> random:Splittable_random.t -> 'a) -> 'a t @@ portable

Creates a generator that calls the given function with the current size parameter and pseudorandom state.

val generate : 'a. 'a t -> size:Base.int -> random:Splittable_random.t -> 'a @@ portable

Generates a random value using the given size and pseudorandom state. Useful when using create and dispatching to other existing generators.

module Via_thunk : sig ... end

module Debug : sig ... end