Module Field_syntax.Record

type ast = Ppx_quickcheck_expander__.Import.label_declaration

type t

val create : ast -> t

val location : t -> Ppx_quickcheck_expander__.Import.location

location of the field declaration

val core_type : t -> Ppx_quickcheck_expander__.Import.core_type

type of the field's contents

val pattern : t Ppx_quickcheck_expander__.Import.list -> loc:Ppx_quickcheck_expander__.Import.location -> Ppx_quickcheck_expander__.Import.pattern Ppx_quickcheck_expander__.Import.list -> Ppx_quickcheck_expander__.Import.pattern

constructing a pattern to match all fields of the type

val expression : t Ppx_quickcheck_expander__.Import.list -> loc:Ppx_quickcheck_expander__.Import.location -> Ppx_quickcheck_expander__.Import.expression Ppx_quickcheck_expander__.Import.list -> Ppx_quickcheck_expander__.Import.expression

constructing an expression filling in all fields of the type