Module Test.Config

module Seed : sig ... end

type t = {

seed : Seed.t;	(*
seed is used to initialize the pseudo-random state before running tests of a property.
*)

test_count : Base.int;	(*
test_count determines how many random values to test a property with.
*)

shrink_count : Base.int;	(*
shrink_count determines the maximum number of attempts to find a smaller version of a value that fails a test.
*)

sizes : Base.int Base.Sequence.t;	(*
sizes determines the progression of value sizes to generate while testing. Testing fails if sizes is not of length at least test_count.
*)

}

val sizes : t -> Base.int Base.Sequence.t

val shrink_count : t -> Base.int

val test_count : t -> Base.int

val seed : t -> Seed.t

module Fields : sig ... end

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t